Initial commit of espmcpdaq
This is very simple ADC board developed using ESP32 and MCP3208.
Up to 4 MCPs can be used with a total of 32 single ended channels.

It uses LM336-2.5V as a reference voltage for ADC. This repository
also contains the arduino files for the ESP and a python interface,
server and client.